+++ Source/ARX/AR/include/ARX/AR/ar.h
@@ -1657,27 +1657,44 @@ typedef enum {
         Change to the root of the implementation-dependent user-writable root.
         On iOS and sandboxed macOS, this is equivalent to the root of the application sandbox.
         On Linux and non-sandboxed macOS, this is equivalent to the "~" user home.
-        On Android, this is the root of the "external" storage (e.g. an SD card).
+        On Android, this is the root of the "external" storage (e.g. an SD card), and read access
+        requires READ_EXTERNAL_STORAGE or WRITE_EXTERNAL_STORAGE permission. Note that as of
+        Android 10 (API level 29) this is no longer writable by applications.
         On Windows, this is the user home directory, typically "C:\Documents and Settings\USERNAME" or "C:\Users\USERNAME".
         On Windows UWP
      */
     AR_UTIL_RESOURCES_DIRECTORY_BEHAVIOR_USE_USER_ROOT,
     /*!
         Change to a writable cache directory, i.e. a directory which is not normally shown to the user, in which files which may be subject to deletion by the system or the user.
-        On Android, this is the applications's (internal) cache directory, and a valid instance of Android/Context must be passed in the instanceofAndroidContext parameter.
+        On Android, this is the applications's (internal) cache directory, and a valid instance of Android/Context must be passed in the instanceofAndroidContext parameter,
+        or if none is passed, an attempt will be made using the ApplicationContext.
      */
     AR_UTIL_RESOURCES_DIRECTORY_BEHAVIOR_USE_APP_CACHE_DIR,
     /*!
         Change to a writable data directory, i.e. a directory which is not normally shown to the user, but in which files are retained permanently.
-        On Android, this is the applications's (internal) files directory, and a valid instance of Android/Context must be passed in the instanceofAndroidContext parameter.
+        On Android, this is the applications's (internal) files directory, and a valid instance of Android/Context must be passed in the instanceofAndroidContext parameter,
+        or if none is passed, an attempt will be made using the ApplicationContext.
      */
 	AR_UTIL_RESOURCES_DIRECTORY_BEHAVIOR_USE_APP_DATA_DIR,
     /*!
      Change to a writable temporary directory, i.e. a directory which is not normally shown to the user, and from which files may be deleted at the end of program execution.
-     On Android, this is the applications's (internal) cache directory, and a valid instance of Android/Context must be passed in the instanceofAndroidContext parameter.
+     On Android, this is the applications's (internal) cache directory, and a valid instance of Android/Context must be passed in the instanceofAndroidContext parameter,
+        or if none is passed, an attempt will be made using the ApplicationContext.
      */
-    AR_UTIL_RESOURCES_DIRECTORY_BEHAVIOR_USE_TMP_DIR
+    AR_UTIL_RESOURCES_DIRECTORY_BEHAVIOR_USE_TMP_DIR,
+    /*!
+        Change to a writable data directory, i.e. a directory which is not normally shown to the user, but in which files are retained permanently.
+        On Android, this is the applications's (external) files directory, and a valid instance of Android/Context must be passed in the instanceofAndroidContext parameter,
+        or if none is passed, an attempt will be made using the ApplicationContext.
+        On other platforms, this currently has the same behaviour as AR_UTIL_RESOURCES_DIRECTORY_BEHAVIOR_USE_APP_DATA_DIR.
+     */
+	AR_UTIL_RESOURCES_DIRECTORY_BEHAVIOR_USE_APP_EXTERNAL_DATA_DIR
+
 } AR_UTIL_RESOURCES_DIRECTORY_BEHAVIOR;
+
+#ifdef ANDROID
+jobject arUtilGetGlobalContext(void);
+#endif
 
 /*!
     @brief   Get the path to the resources directory using the specified behavior.